requirements.txt
openai==0.28
streamlit
bert-extractive-summarizer
torch
sentence-transformers
google-generativeai
pyyaml
